chore: add reusable function to common
